move resource manager interface into plugin core
